chore: converting util function to separate script
