+++ src/maud/cli.py
@@ -26,7 +26,7 @@ import toml
 from maud import sampling
 from maud.analysis import load_infd
 from maud.io import load_maud_input_from_toml, parse_config, parse_toml_kinetic_model
-from maud.utils import get_input_template
+from maud.get_prior_template import get_prior_template
 
 
 RELATIVE_PATH_EXAMPLE = "../../tests/data/linear"
@@ -142,7 +142,7 @@ def generate_prior_template(data_path):
     output_name = "prior_template.csv"
     output_path = os.path.join(data_path, output_name)
     print("Creating template: " + output_path)
-    prior_dataframe = get_input_template(kinetic_model, raw_measurements)
+    prior_dataframe = get_prior_template(kinetic_model, raw_measurements)
     print(f"Saving template to: " + output_path)
     prior_dataframe.to_csv(output_path)
     print("Generated CSV")
